refactor(providers): engine de estratégias de captura (interceptor/content)
Substitui o if-cascade por provider em src/interceptor/index.ts (MAIN) e
src/content/index.ts (ISOLATED) por um catálogo de estratégias declaradas por
provider (src/capture/strategies.ts) e um registry (src/capture/registry.ts).

Cada provider declara COMO captura num módulo capture.ts separado do index.ts de
processamento, com a lógica EXATA movida (sem reescrita):
- x/capture.ts: networkIntercept (extractXEndpointName).
- linkedin/capture.ts: networkIntercept (queryId) + embeddedCodeScan (bpr).
- instagram/capture.ts: networkIntercept (markers/gate + rename) + ssrScriptScan
  + liveDomScrapes (VISIBLE_PUBLICATIONS/VISIBLE_COMMENTS).

Os motores viram engines genéricos que iteram o registry e rodam as estratégias do
provider ativo. O protocolo de mensagens (src/shared/messages.ts) NÃO muda: as
mensagens emitidas (CAPTURED_PAYLOAD/VISIBLE_PUBLICATIONS/VISIBLE_COMMENTS) e o
agendamento (timers/observers) são idênticos aos de antes. controller.ts, store.ts e
providers/*/index.ts de processamento não foram tocados.

Camada sem testes automatizados: validada por typecheck (3 erros pré-existentes,
inalterado) e build. PRECISA de verificação no browser (X/IG/LinkedIn).

Refs #5 #7
